Redid find method
Previously, to get all carnivores (insectivores and piscivores
included), user would input a "targets" array that included
["Carnivore", "Insectivore", "Piscivore"], and then the
Dinodex.find method would look for all dinos that had a
diet matching any of the elements in the targets array.

After discussing with Parul and Jordan, decided that the code
required to do the above was too convoluted/unreadable.
Instead, chose to treat a search for "carnivore" as a special
case (much like a search based on weight) such that the find
method would also return dinosaurs with insectivore/piscivore
diets along with carnivores.

This removes the need to have a targets array; instead, the
find method only accepts a single target.