Refactored dinodex to raise an error
Raises error upon initialization if one of the input CSV files
does not match an accepted format.

Also refactored a few predicate methods to make sure that they
do return a boolean method, or nil in the case of matches_weight?
(indicating an error -- that user tried to search by a weight
that wasn't 'big' or 'small'. still trying to figure out how/when
to catch this error)